requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
google-generativeai>=0.5.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
pillow==10.2.0
python-magic==0.4.27
httpx==0.27.0
pytest==8.1.1
pytest-asyncio==0.23.5
loguru>=0.7.0
